Create function that calculates total input prices
The function takes the user input data already calculated and calculates
the total price. The function also prints out this data.

-There is currently a bug involving the salePrice of items not working
correctly. If a number of the item not divisible by the sale
number for that item is used as an input, the resulting price total
will be too high.